Add files via upload
Libraries:
pandas numpy matplotlib sklearn


Project motivation
This program would tell which things affect your placement? Sounds great, right?
Let's look at how we can accomplish this with the help of Data Analysis! 
and predict the Percentage of Placement campus with the help of Machine Learning!
Files in the repository

dataset/Placement_Data_Full_Class.csv : The dataset file.
placement_analysis.ipynb : Notebook file contaning the exploration and analysis of the work.
readme.md : The readme file for write-up.


Dataset
What is in it?
This data set consists of Placement data of students in Jain University Bangalore.  It includes secondary and higher secondary school percentage and specialization. It also includes degree specialization, type and Work experience and salary offers to the placed students.

Following are the objectives of the project:

Overview of the Problem… Questions

Which factor influenced a candidate in getting placed?

or

Does percentage matters for one to get placed?

what is the percentage of female or male how get a placement?

Which degree specialization is much demanded by corporate?

Predictiction model using differnt ML algorithms?
To classify whether student will get placed in any company or not. To study the various classification models. To compare the accuracies between the various classification models. To study the performance measures. Following step are performed:

1.Gathering Data 2.Preparing that Data 3.Data visualization and cleaning 4.Training 5.Evaluation 6.Prediction In this project, I have used two classifiers :

1.Logistic Regression 2.Random Forest Classifier

Summary of the results
result result result result result result

Acknowledgements
kaggle Providing the Placement datasets.

stackoverflow